backend/pyproject.toml
[project]
name = "jailmaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.5.0",
    "django-ninja>=1.3.0",
    "gspread>=6.1.3",
    "ortools>=9.11.4210",
    "pandas[excel]>=2.2.3",
    "pypdf>=5.1.0",
    "ruff>=0.6.9",
]

[tool.ruff]
line-length = 120
lint.select = [
    "B",    # flake8-bugbear
    "C",    # flake8-comprehensions
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "FURB", # refurb
    "I",    # isort
    "PTH",  # flake8-use-pathlib
    "UP",   # pyupgrade
    "W",    # pycodestyle warnings
]
